Logic is broken in main_bu_code
The logic of the test is broken. "retcode" can not be equals to a String and then equals  to 0 in the block of code that follows.
Either an "else:" is missing either the indentation of the block is wrong.
Note that I doubt that retcode will ever get a String value.